修改cartable以支援max_rpm到更大數值單位 新增seeder測試內容

+++ database/migrations/2023_10_18_013216_create_car_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->tinyInteger('riding_noise')->unsigned()->comment('騎乘噪音值');
             $table->tinyInteger('idle_noise')->unsigned()->comment("怠速噪音值");
             $table->double('max_power')->unsigned()->comment("最大動力");
-            $table->tinyInteger('max_rpm')->unsigned()->comment("最大動力轉速");
+            $table->unsignedInteger('max_rpm')->unsigned()->comment("最大動力轉速");
             $table->double('displacement')->unsigned()->comment("排氣量");
             $table->timestamps();
